soft code finding of suggestions
